Adjust some variables (mostly related to the buffer cache) that hold address space sizes to be longs instead of ints.  Specifically, the follow values are now longs: runningbufspace, bufspace, maxbufspace, bufmallocspace, maxbufmallocspace, lobufspace, hibufspace, lorunningspace, hirunningspace, maxswzone, maxbcache, and maxpipekva.  Previously, a relatively small number (~ 44000) of buffers set in kern.nbuf would result in integer overflows resulting either in hangs or bogus values of hidirtybuffers and lodirtybuffers.  Now one has to overflow a long to see such problems.  There was a check for a nbuf setting that would cause overflows in the auto-tuning of nbuf.  I've changed it to always check and cap nbuf but warn if a user-supplied tunable would cause overflow.
Note that this changes the ABI of several sysctls that are used by things
like top(1), etc., so any MFC would probably require a some gross shims
to allow for that.

MFC after:	1 month

+++ sys/sys/pipe.h
@@ -56,7 +56,7 @@
 /*
  * See sys_pipe.c for info on what these limits mean. 
  */
-extern int	maxpipekva;
+extern u_long	maxpipekva;
 
 /*
  * Pipe buffer information.
